+++ crates/interpreter/src/lib.rs
@@ -106,13 +106,13 @@
 
 #![cfg_attr(not(feature = "debug"), no_std)]
 #![cfg_attr(not(feature = "toctou"), feature(slice_split_at_unchecked))]
+#![cfg_attr(test, allow(unused_crate_dependencies))]
 #![feature(concat_idents)]
 #![feature(float_minimum_maximum)]
 #![feature(never_type)]
 #![feature(pointer_is_aligned)]
 #![feature(try_blocks)]
 #![feature(unwrap_infallible)]
-#![warn(unsafe_op_in_unsafe_fn)]
 
 extern crate alloc;