+++ crates/scheduler/src/call.rs
@@ -23,7 +23,7 @@ macro_rules! or_trap {
         #[cfg(feature = $feature)]
         $name($call);
         #[cfg(not(feature = $feature))]
-        $call.reply_(Err(crate::Trap));
+        $call.reply_(Err(crate::Trap.into()));
     }};
 }
 
@@ -33,7 +33,9 @@ macro_rules! or_fail {
         #[cfg(feature = $feature)]
         $name($call);
         #[cfg(not(feature = $feature))]
-        $call.reply_(Ok(Err(wasefire_error::Error::world(wasefire_error::Code::NotImplemented))));
+        $call.reply_(Err(crate::Failure::Error(wasefire_error::Error::world(
+            wasefire_error::Code::NotImplemented,
+        ))));
     }};
 }
 
@@ -113,5 +115,5 @@ pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
 
 fn syscall<B: Board>(call: SchedulerCall<B, api::syscall::Sig>) {
     let api::syscall::Params { x1, x2, x3, x4 } = call.read();
-    call.reply(B::syscall(*x1, *x2, *x3, *x4).ok_or(Trap));
+    call.reply(try { B::syscall(*x1, *x2, *x3, *x4).ok_or(Trap)?? });
 }