+++ crates/api-macro/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Patch
 
+- Use common Wasefire lints
 - Return a user error instead of an internal error for enums `TryFrom`
 
 ## 0.6.0